llm-backend/requirements.txt
pymongo==4.4.1
beanie==1.20.0
pydantic~=1.10.11
fastapi==0.100.1
python-dateutil==2.8.2
uvicorn
openai==1.3.6
python-dotenv